refactor: use pascalcase in NoTaskFound component folder
