+++ src/components/NoTaskFound/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Image from "components/Common/Image";
-import styles from "components/not-found/task/index.module.scss";
+import styles from "components/NoTaskFound/index.module.scss";
 import {
   ALT_TAG_NO_TASK_FOUND_ICON,
   ICON_NO_TASK_FOUND,
